refactor: add internal
