+++ pkg/app/server/binding/default.go
@@ -68,7 +68,7 @@ import (
 
 	"github.com/bytedance/go-tagexpr/v2/validator"
 	"github.com/cloudwego/hertz/internal/bytesconv"
-	"github.com/cloudwego/hertz/pkg/app/server/binding/decoder"
+	"github.com/cloudwego/hertz/pkg/app/server/binding/internal/decoder"
 	hjson "github.com/cloudwego/hertz/pkg/common/json"
 	"github.com/cloudwego/hertz/pkg/common/utils"
 	"github.com/cloudwego/hertz/pkg/protocol"
