+++ pkg/app/server/binding/internal/decoder/map_type_decoder.go
