+++ pkg/app/server/binding/internal/decoder/slice_getter.go
