+++ pkg/app/server/binding/internal/decoder/struct_type_decoder.go
