+++ pkg/app/server/binding/internal/decoder/text_decoder.go
